+++ src/middleware.ts
@@ -1,0 +1,22 @@
+import { NextResponse } from "next/server";
+import type { NextRequest } from "next/server";
+
+export async function middleware(request: NextRequest) {
+  const token = request.cookies.get("token")?.value;
+  const { pathname } = request.nextUrl;
+
+  const protectedURL = ["/myPage", "/home"];
+  if (!token && protectedURL.some((route) => pathname.startsWith(route)))
+    return NextResponse.redirect(new URL("/login", request.url));
+
+  const publicURL = ["/login", "/signUp", "/profile"];
+  if (token && publicURL.some((route) => pathname.startsWith(route)))
+    return NextResponse.redirect(new URL("/home", request.url));
+
+  return NextResponse.next();
+}
+
+// matcher 설정
+export const config = {
+  matcher: ["/myPage", "/home", "/login", "/signUp", "/profile"],
+};
